feat: Use deployment instead of statefulset for workers (#10)
* feat: Use deployment instead of statefulset for workers

Polars on-premises currently does not consider locally mounted storage as persistent, and probably won't for the next while. For shuffles we may support using S3 instead, as a persistent shared storage. This changed requirement means we don't have the need for a statefulset anymore. A deployment is quite a bit easier to work with, so here we go.

* remove irrelevant example